% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-themes.R
\name{bulma_themes}
\alias{bulma_themes}
\title{Themes}
\usage{
bulma_themes()
}
\description{
List all available themes.
}
\details{
All themes are taken from \href{https://jenil.github.io/bulmaswatch/}{bulmaswatch}.
}
