% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-input.R
\name{bulmaActionButton}
\alias{bulmaActionButton}
\title{Add action button}
\usage{
bulmaActionButton(inputId, label, color = NULL)
}
\arguments{
\item{inputId}{the input slot that will be used to access the value.}

\item{label}{button label.}

\item{color}{button color.}
}
\description{
Add an action button.
}
\examples{
if(interactive()){
library(shiny)

shinyApp(
  ui = bulmaPage(
    bulmaActionButton("goButton", "Render in console")
  ),
  server = function(input, output) {
    observeEvent(input$goButton, {
      print("button clicked")
    })
 }
)
}
}
\author{
John Coene, \email{jcoenep@gmail.com}
}
