% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-alert.R
\name{bulmaAlert}
\alias{bulmaAlert}
\title{Create a bulma JS alert}
\usage{
bulmaAlert(options, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{options}{List containing alert configuration. See \url{https://bulmajs.tomerbe.co.uk/}.}

\item{session}{Shiny session object. Internal use.}
}
\value{
An alert message
}
\description{
Alert extension for bulma
}
\examples{
if (interactive()) {
 library(shiny)
 library(shinybulma)
 
 ui <- bulmaPage(
   bulmaActionButton("alert", "Show alert")
 )
 
 server <- function(input, output, session) {
   observeEvent(input$alert, {
     bulmaAlert(
       list(
        type = "warning",
        title = "Warning!!!", 
        body = "Alert content", 
        confirm = "Ok", 
        cancel = "cancel",
        showHeader = TRUE
       )
     )
   })
 }
 
 shinyApp(ui, server)
}
}
