% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-badge.R
\name{bulmaBadge}
\alias{bulmaBadge}
\title{bulma badge components}
\usage{
bulmaBadge(..., color = NULL, size = NULL, style = NULL,
  position = NULL, badge_text = NULL)
}
\arguments{
\item{...}{Any element where the badge will be stuck.}

\item{color}{Badge color : \code{link}, \code{info}, \code{primary}, \code{warning},
\code{danger}, \code{success}, \code{black}, \code{dark} and \code{ligth}.}

\item{size}{Badge size : \code{small}, \code{medium} and \code{large}.}

\item{style}{Badge style :  \code{outlined}. NULL by default}

\item{position}{Badge position : \code{left}, \code{bottom-left} and \code{bottom}. When NULL, it is displayed on the right side by default.}

\item{badge_text}{Badge content : avoid too large content (a number is better).}
}
\description{
Create a badge to highlight a given element \url{https://wikiki.github.io/elements/badge/}.
}
\examples{
if (interactive()) {
 library(shiny)

 ui <- bulmaPage(
 bulmaContainer(
 br(), br(),
 bulmaColumns(
  bulmaColumn(
   width = 2,
   bulmaBadge(
    "test",
    position = "",
    size = NULL,
    style = NULL,
    color = NULL,
    badge_text = 3
   ),

   br(), br(),

   bulmaBadge(
     "test",
     position = "bottom",
     size = NULL,
     style = NULL,
     color = NULL,
     badge_text = 3
   ),

   br(), br(),

   bulmaBadge(
     "test",
     position = "bottom-left",
     size = NULL,
     style = NULL,
     color = NULL,
     badge_text = 3
   )
   ),
   bulmaColumn(
     width = 2,
     bulmaBadge(
       "test",
       position = NULL,
       size = "small",
       style = NULL,
       color = NULL,
       badge_text = 3
     ),

     br(), br(),

     bulmaBadge(
       "test",
       position = NULL,
       size = "medium",
       style = NULL,
       color = NULL,
       badge_text = 3
     ),

     br(), br(),

     bulmaBadge(
       "test",
       position = NULL,
       size = "large",
       style = NULL,
       color = NULL,
       badge_text = 3
     )
   ),
   bulmaColumn(
     width = 2,
     bulmaBadge(
       "test",
       position = NULL,
       size = NULL,
       style = "outlined",
       color = "info",
       badge_text = 3
     ),

     br(), br(),

     bulmaBadge(
       "test",
       position = NULL,
       size = NULL,
       style = NULL,
       color = "warning",
       badge_text = 3
     ),

     br(), br(),

     bulmaBadge(
       "test",
       position = NULL,
       size = NULL,
       style = NULL,
       color = "danger",
       badge_text = 3
     ),

     br(), br(),

     bulmaBadge(
       "test",
       position = NULL,
       size = NULL,
       style = NULL,
       color = "success",
       badge_text = 3
     )
    )
   )
  )
 )

 server <- function(input, output, session) {

 }

 shinyApp(ui = ui, server = server)
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
