% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-box.R
\name{bulmaBox}
\alias{bulmaBox}
\title{bulma Box}
\usage{
bulmaBox(...)
}
\arguments{
\item{...}{Any element.}
}
\description{
Create a simple box with or without media \url{https://bulma.io/documentation/elements/box/}.
}
\note{
It is lighter than a bulmaCard element.
}
\examples{
\dontrun{
if (interactive()) {
 library(shiny)

 ui <- bulmaPage(
  bulmaContainer(
   bulmaColumns(
    bulmaColumn(
      width = 6,
      bulmaBox(
        bulmaBoxMedia(
          src = "https://bulma.io/images/placeholders/128x128.png",
          "test"
        )
      )
    ),
    bulmaDivider(vertical = TRUE),
    bulmaColumn(
      width = 6,
      bulmaBox(
        "test"
      )
     )
    )
  )
 )

 server <- function(input, output, session) {

 }

 shinyApp(ui = ui, server = server)

}
}
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
