% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-box.R
\name{bulmaBoxMedia}
\alias{bulmaBoxMedia}
\title{bulma Box media.}
\usage{
bulmaBoxMedia(..., src = NULL, size = "64x64")
}
\arguments{
\item{...}{any element.}

\item{src}{link to the media.}

\item{size}{Media size.}
}
\description{
Create a media to insert in a bulmaBox.
}
\details{
Valid \code{size}:
\itemize{
  \item{16 - 16x16}
  \item{24 - 24x24}
  \item{32 - 32x32}
  \item{48 - 48x48}
  \item{54 - 54x54}
  \item{96 - 96x96}
  \item{128 - 128x128}
}
}
