% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulmaBtn.R
\name{BulmaBtnGroup}
\alias{BulmaBtnGroup}
\title{Bulma Button Group}
\usage{
BulmaBtnGroup(inputId, ...)
}
\arguments{
\item{inputId}{string value for id}

\item{...}{BulmaBtn function}
}
\description{
Bulma Button Group
}
\examples{
BulmaBtnGroup(
 inputId = "example",
 BulmaBtn(inputId = "1", text = "1"),
 BulmaBtn(inputId = "2", text = "2"))
}
