% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-button.R
\name{bulmaButton}
\alias{bulmaButton}
\title{Button}
\usage{
bulmaButton(..., color = NULL, style = NULL, class = NULL,
  rounded = FALSE, size = NULL, display = NULL,
  tag = shiny::tags$a)
}
\arguments{
\item{...}{button content.}

\item{color}{a valid bulma color, i.e.: \code{white}, or \code{primary}.}

\item{style}{can be set to either \code{interted} or \code{outlined}.}

\item{class}{additional CSS classes.}

\item{rounded}{set to \code{TRUE} for a rounded button.}

\item{size}{size of button, \code{small}, \code{medium}, or \code{large}.}

\item{display}{can be set to \code{fullwidth}.}

\item{tag}{shiny tag to use.}
}
\description{
Create a bulma button.
}
\examples{
if(interactive()){
  library(shiny)
  
  ui <- bulmaPage(
    bulmaContainer(
      bulmaButton(
        "Button",
        color = "info",
        rounded = TRUE
      ),
      bulmaButton(
        "Outline",
        style = "outlined"
      )
    )
  )
  
  server <- function(input, output){}
  
  shinyApp(ui, server)
}

}
