% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-buttonBadge.R
\name{bulmaButtonBadge}
\alias{bulmaButtonBadge}
\title{bulma Button Badge}
\usage{
bulmaButtonBadge(label = NULL, url = NULL, color = NULL, style = NULL,
  loading = FALSE, size = NULL, badge_color = NULL, badge_style = NULL,
  badge_size = NULL, badge_position = NULL, badge_text = NULL)
}
\arguments{
\item{label}{button label.}

\item{url}{The url of the button.}

\item{color}{Color of the button : \code{link}, \code{info}, \code{primary}, \code{warning},
\code{danger}, \code{success}, \code{black}, \code{dark} and \code{ligth}.}

\item{style}{Style of the button :  \code{outlined}.}

\item{loading}{if the button is loading or not.}

\item{size}{Size of the button : \code{small},\code{medium} and \code{large}.}

\item{badge_color}{Color of the button : \code{link}, \code{info}, \code{primary}, \code{warning},
\code{danger}, \code{success}, \code{black}, \code{dark} and \code{ligth}.}

\item{badge_style}{Style of the button :  \code{outlined}.}

\item{badge_size}{Size of the button : \code{small},\code{medium} and \code{large}.}

\item{badge_position}{Position of the badge : \code{left}, \code{bottom-left} and \code{bottom}. Right by default.}

\item{badge_text}{Badge content : avoid too large content (a number is better).}
}
\description{
Create a button with a badge \url{https://wikiki.github.io/elements/badge/}.
}
\examples{
\dontrun{
if (interactive()) {
 library(shiny)

 ui <- bulmaPage(
  bulmaContainer(
   bulmaColumns(
   bulmaColumn(
    width = 2,
    bulmaButtonBadge(label = "test",
                     url = "http://google.com", color = "danger", style = NULL, loading = FALSE,
                     size = NULL, badge_color = "warning", badge_style = NULL,
                     badge_size = NULL, badge_position = NULL, badge_text = 1
    ),
    br(), br(),

    bulmaButtonBadge(label = "test",
                     url = NULL, color = "warning", style = "outlined", loading = FALSE,
                     size = "small", badge_color = "info", badge_style = NULL,
                     badge_size = NULL, badge_position = NULL, badge_text = 2
    ),
    br(), br(),

    bulmaButtonBadge(label = "test",
                     url = NULL, color = "success", style = NULL, loading = FALSE,
                     size = "medium", badge_color = "danger", badge_style = NULL,
                     badge_size = NULL, badge_position = NULL, badge_text = 3
    )
    ),
    bulmaColumn(
      width = 2,
      bulmaButtonBadge(label = "test",
                       url = NULL, color = "black", style = "outlined", loading = FALSE,
                       size = "large", badge_color = NULL, badge_style = NULL,
                       badge_size = NULL, badge_position = NULL, badge_text = 4
      ),
      br(), br(),

      bulmaButtonBadge(label = "test",
                       url = NULL, color = "info", style = NULL, loading = FALSE,
                       size = NULL, badge_color = "success", badge_style = NULL,
                       badge_size = "small", badge_position = NULL, badge_text = 5
      ),
      br(), br(),

      bulmaButtonBadge(label = "test",
                       url = NULL, color = NULL, style = "outlined", loading = FALSE,
                       size = NULL, badge_color = "warning", badge_style = NULL,
                       badge_size = "medium", badge_position = "left", badge_text = 6
      )
    ),
    bulmaColumn(
      width = 2,
      bulmaButtonBadge(label = "test",
                       url = NULL, color = NULL, style = "primary", loading = TRUE,
                       size = NULL, badge_color = NULL, badge_style = "outlined",
                       badge_size = "large", badge_position = "botton", badge_text = NULL
      )
    )
   )
  )
 )

 server <- function(input, output, session) {

 }

 shinyApp(ui = ui, server = server)

}
}
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
