% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-carousel.R
\name{bulmaCarousel}
\alias{bulmaCarousel}
\title{bulma carousel}
\usage{
bulmaCarousel(..., autoplay = FALSE, navigation = NULL)
}
\arguments{
\item{...}{Slot for bulmaCarouselItem.}

\item{autoplay}{FALSE by default. If TRUE, the carousel automatically change its content.}

\item{navigation}{Where to display the navigations arrow : \code{centered} or \code{overlay}.
If NULL, arrows are at the bottom and more spaced.}
}
\description{
Create a carousel \url{https://wikiki.github.io/components/carousel/}.
}
\examples{
\dontrun{
if (interactive()) {
 library(shiny)

 ui <- bulmaPage(
  bulmaContainer(
    bulmaCarousel(
     autoplay = TRUE, navigation = "overlay",

     bulmaCarouselItem(
       background = TRUE, active = TRUE,
       src = "https://wikiki.github.io/images/merry-christmas.jpg",
       title = "Merry Christmas", url = NULL, url_text = NULL
     ),
     bulmaCarouselItem(
       background = TRUE, active = FALSE,
       src = "https://wikiki.github.io/images/singer.jpg",
       title = "Original Gift: Offer a song with",
       url = "https://lasongbox.com", url_text = "La Song Box"
     ),
     bulmaCarouselItem(
       background = TRUE, active = FALSE,
       src = "https://wikiki.github.io/images/sushi.jpg",
       title = "Sushi time", url = NULL, url_text = NULL
     ),
     bulmaCarouselItem(
       background = TRUE, active = FALSE,
       src = "https://wikiki.github.io/images/life.jpg",
       title = "Life", url = NULL, url_text = NULL
     )
   )
  )
 )

 server <- function(input, output, session) {

 }

 shinyApp(ui = ui, server = server)

}
}
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
