% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-carousel.R
\name{bulmaCarouselItem}
\alias{bulmaCarouselItem}
\title{bulma carousel item}
\usage{
bulmaCarouselItem(
  background = TRUE,
  active = FALSE,
  src = NULL,
  title = NULL,
  url = NULL,
  url_text = NULL
)
}
\arguments{
\item{background}{by default TRUE. The item background.}

\item{active}{TRUE of FALSE. The item which will be displayed first has to
be TRUE. All others will be FALSE.}

\item{src}{Image path.}

\item{title}{Item title.}

\item{url}{External link, if any.}

\item{url_text}{External link text, if any.}
}
\description{
Create a carousel item \url{https://wikiki.github.io/components/carousel/}.
}
