% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-input.R
\name{bulmaCheckInput}
\alias{bulmaCheckInput}
\title{Check boxes}
\usage{
bulmaCheckInput(inputId, choices, color = NULL, checked = NULL)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value.}

\item{choices}{Vector of choices, named or unnamed.}

\item{color}{Vector of colors.}

\item{checked}{Values of checked boxes.}
}
\description{
Add add check boxes.
}
\examples{
if(interactive()){
shiny::shinyApp(
  ui = bulmaPage(
   bulmaTitle("Hello Bulma"),
   bulmaCheckInput("select", c("Miles per galon" = "mpg", "Rear axle ratio" = "drat"),
     checked = "mpg", color = c("danger", "info")),
   verbatimTextOutput("selected")
  ),
  server = function(input, output) {
    output$selected <- renderText({
      input$select
    })
  }
)
}

}
\author{
John Coene, \email{jcoenep@gmail.com}
}
