% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-column.R
\name{bulmaColumn}
\alias{bulmaColumn}
\title{Add a column}
\usage{
bulmaColumn(..., width = NULL, narrow = FALSE)
}
\arguments{
\item{...}{any element.}

\item{width}{width of column (12-column system).}

\item{narrow}{use for column to only take the space it needs.}
}
\description{
Add a column
}
