% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-column.R
\name{bulmaColumns}
\alias{bulmaColumns}
\title{Add a column}
\usage{
bulmaColumns(..., gapless = FALSE, centered = FALSE, multiline = FALSE,
  mobile = FALSE, tablet = FALSE, desktop = FALSE)
}
\arguments{
\item{...}{any element.}

\item{gapless}{if \code{TRUE} remove gaps between columns.}

\item{centered}{to center columns.}

\item{multiline}{to create a flexible, centered list.}

\item{mobile, tablet, desktop}{set to force columns on device.}
}
\description{
Add a column
}
