% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-input.R
\name{bulmaDateInput}
\alias{bulmaDateInput}
\title{Date picker}
\usage{
bulmaDateInput(inputId, value, min = Sys.Date() - 3, max = Sys.Date(),
  overlay = FALSE, lang = "en", ...)
}
\description{
Add date picker
}
\examples{
if(interactive()){
  library(shiny)
  
  ui <- bulmaPage(
    bulmaContainer(
      br(),
      bulmaDateInput("date", Sys.Date(), overlay = TRUE),
      verbatimTextOutput("selected")
    )
  )
  
  server <- function(input, output){
    
    output$selected <- renderText({
      input$date
    })
  }
  
  shinyApp(ui, server)
}

}
