% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-divider.R
\name{bulmaDivider}
\alias{bulmaDivider}
\title{bulma divider}
\usage{
bulmaDivider(content = NULL, vertical = FALSE)
}
\arguments{
\item{content}{Content to display (In mathematical textbooks we usually find ***).}

\item{vertical}{TRUE or FALSE. Whether to use a vertical or horizontal divider.}
}
\description{
Create a vertical or horizontal divider \url{https://wikiki.github.io/layout/divider/}.
}
\note{
If vertical is TRUE, you need to wrap the divider in a bulmaColumns() configuration. See example below.
}
\examples{
\dontrun{
if (interactive()) {
 library(shiny)

 ui <- bulmaPage(
  bulmaContainer(
   # horizontal divider
   bulmaDivider(),
   # vertical divider need to be wrapped in columns!
   bulmaColumns(
    bulmaColumn(
      width = 6,
      bulmaBox(
         "test"
      )
    ),
    bulmaDivider(vertical = TRUE),
    bulmaColumn(
      width = 6,
      bulmaBox(
        "test"
      )
    )
   )
  )
 )

 server <- function(input, output, session) {

 }

 shinyApp(ui = ui, server = server)

}
}
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
