% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-charts.R
\name{bulmaFigure}
\alias{bulmaFigure}
\title{Add image}
\usage{
bulmaFigure(src, size = NULL, ratio = NULL, alt = "")
}
\arguments{
\item{src}{path to image.}

\item{size}{set fixed image size, see details.}

\item{ratio}{set responsive ratio, see details.}

\item{alt}{image alt.}
}
\description{
Add an image.
}
\details{
Valid \code{size}:
\itemize{
  \item{16 - 16x16}
  \item{24 - 24x24}
  \item{32 - 32x32}
  \item{48 - 48x48}
  \item{54 - 54x54}
  \item{96 - 96x96}
  \item{128 - 128x128}
}

Valid \code{ratio}:
\itemize{
  \item{square or 1by1}
  \item{1by1}
  \item{4by3}
  \item{3by2}
  \item{16by9}
  \item{2by1}
}
}
\examples{
if(interactive()){
library(shiny)

shinyApp(
  ui = bulmaPage(
   bulmaTitle("Hello Bulma"),
   bulmaFigure("path/to/image")
  ),
  server = function(input, output) {}
)
}

}
\author{
John Coene, \email{jcoenep@gmail.com}
}
