% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-footer.R
\name{bulmaFooter}
\alias{bulmaFooter}
\title{Add a footer}
\usage{
bulmaFooter(..., container = TRUE, centered = TRUE)
}
\arguments{
\item{...}{any element.}

\item{container}{If \code{TRUE} wraps a \code{\link{bulmaContainer}} as a children of the section.}

\item{centered}{wheter to center the text.}
}
\description{
Add footer
}
\examples{
if(interactive()){
library(shiny)

ui <- bulmaPage(
  bulmaHero(
    fullheight = TRUE,
    color = "primary",
    bulmaHeroHead(
      bulmaContainer(
        bulmaTitle("Hero Title")
      )
    ),
    bulmaHeroHead(
      bulmaContainer(
        bulmaSubtitle(
          "Some content"
        )
      )
    ),
    bulmaHeroFoot(
      bulmaContainer(
        "The footer"
      )
    )
  )
)

server <- function(input, output){}

shinyApp(ui, server)
}

}
\author{
John Coene, \email{jcoenep@gmail.com}
}
