% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-icon.R
\name{bulmaIcon}
\alias{bulmaIcon}
\title{Icons}
\usage{
bulmaIcon(icon, color = NULL, size = NULL)
}
\arguments{
\item{icon}{Fontawesome icon class.}

\item{color}{color of icon.}

\item{size}{size of icon.}
}
\description{
Add icons.
}
\examples{
if(interactive()){
  library(shiny)
  
  ui <- bulmaPage(
    bulmaContainer(
      bulmaIcon(
        "fa fa-info",
        color = "info"
      ),
      bulmaButton(
        bulmaIcon("fa fa-home"),
        style = "outlined"
      )
    )
  )
  
  server <- function(input, output){}
  
  shinyApp(ui, server)
}

}
