% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-loading.R
\name{bulmaLoading}
\alias{bulmaLoading}
\title{Show loading page}
\usage{
bulmaLoading(title = "Loading", class = NULL)
}
\arguments{
\item{title}{Title of loading page.}

\item{class}{Any additional class.}
}
\description{
Show loading page
}
\examples{
if(interactive()){
  ui <- bulmaPage(
    bulmaTitle("Loader"),
    bulmaSliderInput("slider", 50, 5, 100, step = 5),
    bulmaLoading(),
    bulmaContainer(
      plotOutput("plot")
    )
  )
  
  server <- function(input, output){
    data <- reactive({
      rnorm(input$slider, 10, 5)
    })
    
    output$plot <- renderPlot({
      hist(data())
    })
  }
  
  shiny::shinyApp(ui, server)
}

}
\author{
John Coene, \email{jcoenep@gmail.com}
}
