% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-menu.R
\name{bulmaMenuItem}
\alias{bulmaMenuItem}
\title{bulma menu item}
\usage{
bulmaMenuItem(..., active = FALSE, item_label = NULL, target = NULL)
}
\arguments{
\item{...}{Slot for bulmaMenuSubItem.}

\item{active}{If the item is active or not. FALSE by default.}

\item{item_label}{Subsection label.}

\item{target}{Target to set an html anchor.}
}
\description{
Create an item to insert in a bulmaMenu object \url{https://bulma.io/documentation/components/menu/}.
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
