% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-menu.R
\name{bulmaMenuLabel}
\alias{bulmaMenuLabel}
\title{bulma menu label}
\usage{
bulmaMenuLabel(menu_label = NULL, target = NULL)
}
\arguments{
\item{menu_label}{Title of the global menu section.}

\item{target}{Target to set an html anchor.}
}
\description{
Create an label to insert in a bulmaMenu object \url{https://bulma.io/documentation/components/menu/}.
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
