% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-menu.R
\name{bulmaMenuSubItem}
\alias{bulmaMenuSubItem}
\title{bulma menu sub item}
\usage{
bulmaMenuSubItem(subitem_label = NULL, target = NULL)
}
\arguments{
\item{subitem_label}{Subsubsection label.}

\item{target}{Target to set an html anchor.}
}
\description{
Create an sub item to insert in q bulmaMenuItem \url{https://bulma.io/documentation/components/menu/}.
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
