% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-modal.R
\name{bulmaModal}
\alias{bulmaModal}
\title{Modal}
\usage{
bulmaModal(title, body, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{title, body}{Content of the modal.}

\item{session}{A valid shiny session.}
}
\description{
Display a modal
}
\examples{
ui <- bulmaPage(
 bulmaActionButton("alert", "Send alert")
)

server <- function(input, output){
 observeEvent(input$alert, {
   bulmaModal("Hello", "World")
 })
}

if(interactive())
 shinyApp(ui, server)

}
