% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-notification.R
\name{bulmaNotification}
\alias{bulmaNotification}
\title{bulma notification}
\usage{
bulmaNotification(
  body,
  color = NULL,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{body}{Content of the alert.}

\item{color}{The background color : \code{link}, \code{info}, \code{primary}, \code{warning},
\code{danger}, \code{success}, \code{black}, \code{dark} and \code{ligth}.}

\item{session}{A valid shiny session.}
}
\description{
Create a notification box \url{https://bulma.io/documentation/elements/notification/}.
}
\examples{
library(shiny)

ui <- bulmaPage(
 bulmaActionButton("alert", "Send alert")
)

server <- function(input, output){
 observeEvent(input$alert, {
   bulmaNotification(color = "red")
 })
}

if(interactive())
 shinyApp(ui, server)

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
