% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-pricingtable.R
\name{bulmaPricingPlan}
\alias{bulmaPricingPlan}
\title{bulma pricing table plan}
\usage{
bulmaPricingPlan(..., active = FALSE, color = NULL, plan_title = NULL,
  plan_price = NULL, plan_currency = NULL, plan_period = NULL,
  button_status = NULL, button_name = NULL)
}
\arguments{
\item{...}{Slot for bulmaPricingPlanItem.}

\item{active}{FALSE by default. If TRUE, the corresponding pricing plan
is highlighted.}

\item{color}{Plan color : \code{link}, \code{info}, \code{primary}, \code{warning},
\code{danger}, \code{success}, \code{black}, \code{dark} and \code{ligth}.}

\item{plan_title}{Plan title.}

\item{plan_price}{Plan price.}

\item{plan_currency}{Plan currency.}

\item{plan_period}{Payment periodicity.}

\item{button_status}{Button status. NULL or \code{"disabled"}.}

\item{button_name}{Button label.}
}
\description{
Create a pricing table plan to insert in a bulmaPricingTable \url{https://wikiki.github.io/components/pricingtable/}.
}
