% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-pricingtable.R
\name{bulmaPricingPlanItem}
\alias{bulmaPricingPlanItem}
\title{bulma pricing plan item}
\usage{
bulmaPricingPlanItem(name = NULL)
}
\arguments{
\item{name}{Item name.}
}
\description{
bulma item to insert in a bulmaPricingPlan container \url{https://wikiki.github.io/components/pricingtable/}.
}
