% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-input.R
\name{bulmaRadioInput}
\alias{bulmaRadioInput}
\title{Add radio buttons}
\usage{
bulmaRadioInput(inputId, choices)
}
\arguments{
\item{inputId}{the input slot that will be used to access the value.}

\item{choices}{vector of choices.}
}
\description{
Add radio buttons
}
\examples{
library(shiny)

shinyApp(
  ui = bulmaPage(
   bulmaTitle("Hello Bulma"),
   bulmaRadioInput("radio-buttons", c("Yes", "No"))
  ),
  server = function(input, output) {}
)

}
