% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-input.R
\name{bulmaSelectInput}
\alias{bulmaSelectInput}
\title{Add dropdown input}
\usage{
bulmaSelectInput(inputId, label, choices)
}
\arguments{
\item{inputId}{the input slot that will be used to access the value.}

\item{label}{input label.}

\item{choices}{vector of choices.}
}
\description{
Add dropdown input
}
\examples{
library(shiny)

shinyApp(
  ui = bulmaPage(
   bulmaTitle("Hello Bulma"),
   bulmaSelectInput("dropdown", "Select", choices = c("Shiny", "Bulma"))
  ),
  server = function(input, output) {}
)

}
