% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-sidebar.R
\name{bulmaSidebar}
\alias{bulmaSidebar}
\title{bulma right sidebar}
\usage{
bulmaSidebar(..., header_title = NULL, footer_content = NULL)
}
\arguments{
\item{...}{Any element to include in the sidebar (such as input elements, ...)}

\item{header_title}{Sidebar header.}

\item{footer_content}{Footer content.}
}
\description{
Create a rigth sidebar \url{https://wikiki.github.io/components/quickview/}.
}
\note{
Not compatible with mobile display.
}
\examples{
if (interactive()) {
 library(shiny)
 
 ui <- bulmaPage(
   bulmaContainer(
   bulmaTitle("Right sidebar demo"),  
   HTML(
       paste0('<button class="button is-info" 
               data-show="quickview" data-target="quickviewDefault">
               <span class="icon"><i class="fas fa-cogs"></i>
               </span></button>
              ')
    )
   ),
   bulmaSidebar(
     header_title = "test",
     
     bulmaSubtitle("Input elements"),
     
     bulmaSliderInput("slider", color = "warning", 10, 3, 150),
     
     bulmaSwitchInput(
       inputId = "switch3", 
       label = "switch 3", 
       value = TRUE,
       
       color = "danger", 
       size = "small", 
       style = "thin",
       
       rtl = FALSE, 
       disabled = FALSE),
     
     bulmaSubtitle("Other Stuff")
     
   ),
   
   plotOutput("plot")
     )
 
 server <- function(input, output, session) {
   data <- reactive({
     rnorm(input$slider, 20, 4)
   })
   
   output$plot <- renderPlot({
     hist(data())
   })
 }
 
 shinyApp(ui = ui, server = server)
}
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
