% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-sidebar.R
\name{bulmaSidebarTrigger}
\alias{bulmaSidebarTrigger}
\title{bulma right sidebar trigger}
\usage{
bulmaSidebarTrigger(color = NULL, size = NULL, style = NULL,
  state = NULL)
}
\arguments{
\item{color}{Button color : \code{link}, \code{info}, \code{primary}, \code{warning},
\code{danger}, \code{success}, \code{black}, \code{dark} and \code{ligth}.}

\item{size}{Button size : \code{small},\code{medium} and \code{large}.}

\item{style}{Button style :  \code{outlined}, \code{inverted}, \code{rounded}}

\item{state}{Button state : \code{hovered} or \code{focused}. NULL by default.}
}
\description{
Create a button to trigger the rigth sidebar \url{https://wikiki.github.io/components/quickview/}.
}
\note{
Can be included in the navbar.
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
