% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-input.R
\name{bulmaSliderInput}
\alias{bulmaSliderInput}
\title{Slider}
\usage{
bulmaSliderInput(value, min, max, step = 1, class = NULL, size = NULL,
  orient = "horizontal")
}
\arguments{
\item{value, min, max}{Current value, maximum and minimum of slider.}

\item{class}{Additional class.}

\item{size}{Size of slider, see details.}

\item{orient}{Slider orientation, takes, \code{horizontal} or \code{vertical}.}
}
\description{
Add a slider input
}
\details{
Valid \code{size}:
\itemize{
  \item{\code{NULL} (standard)}
  \item{\code{small}}
  \item{\code{medium}}
  \item{\code{large}}
}
}
