% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-steps.R
\name{bulmaStepDetail}
\alias{bulmaStepDetail}
\title{bulma step detail}
\usage{
bulmaStepDetail(..., color = NULL, completed = FALSE, title = NULL,
  marker = NULL)
}
\arguments{
\item{...}{Description of the step.}

\item{color}{Step color : \code{link}, \code{info}, \code{primary}, \code{warning},
\code{danger}, \code{success}, \code{black}, \code{dark} and \code{ligth}.}

\item{completed}{TRUE or FALSE. Is the current step completed or not?}

\item{title}{Step title.}

\item{marker}{Whether to put a marker in the round circle.}
}
\description{
Create a step detail to insert in a bulmaStepItem \url{https://wikiki.github.io/components/steps/}.
}
