% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-steps.R
\name{bulmaStepItem}
\alias{bulmaStepItem}
\title{bulma Step item}
\usage{
bulmaStepItem(..., color = NULL, completed = FALSE, active = FALSE)
}
\arguments{
\item{...}{Slot for bulmaStepDetail.}

\item{color}{Plan color : \code{link}, \code{info}, \code{primary}, \code{warning},
\code{danger}, \code{success}, \code{black}, \code{dark} and \code{ligth}.}

\item{completed}{TRUE or FALSE. Is the current step completed or not?}

\item{active}{TRUE or FALSE. Is the current step active or not?}
}
\description{
Create a step item to insert in bulmaSteps \url{https://wikiki.github.io/components/steps/}.
}
