% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-steps.R
\name{bulmaSteps}
\alias{bulmaSteps}
\title{bulma steps}
\usage{
bulmaSteps(..., size = NULL)
}
\arguments{
\item{...}{Slot for bulmaStepItem.}

\item{size}{Global size : \code{small},\code{medium} and \code{large}.}
}
\description{
Create a step element \url{https://wikiki.github.io/components/steps/}.
}
\examples{
if (interactive()) {
 library(shiny)

 ui <- bulmaPage(
  bulmaContainer(
   bulmaSteps(
    size = NULL,
    bulmaStepItem(
      color = "success", completed = TRUE, active = FALSE,
      bulmaStepDetail(
        color = NULL, completed = FALSE, title = "Step 1", marker = 1,
        "This is the first step of the process."
      )
    ),
    bulmaStepItem(
      color = NULL, completed = FALSE, active = FALSE,
      bulmaStepDetail(
        color = NULL, completed = FALSE, title = "Step 2", marker = NULL,
        "This is the second step. You get here once you have completed the first step."
      )
    ),
    bulmaStepItem(
      color = NULL, completed = FALSE, active = FALSE,
      bulmaStepDetail(
        color = NULL, completed = FALSE, title = "Step 3", marker = 3,
        "This is the third step. One more step before the end."
      )
    ),
    bulmaStepItem(
      color = NULL, completed = FALSE, active = TRUE,
      bulmaStepDetail(
        color = NULL, completed = FALSE, title = "Step 4", marker = NULL,
        "Final step. You have completed all the previous steps and end the process."
      )
    )
    )
  )
 )

 server <- function(input, output, session) {

 }

 shinyApp(ui = ui, server = server)

}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
