% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-input.R
\name{bulmaSubmitButton}
\alias{bulmaSubmitButton}
\title{Add submit button}
\usage{
bulmaSubmitButton(inputId, label, color)
}
\arguments{
\item{inputId}{the input slot that will be used to access the value.}

\item{label}{button label.}

\item{color}{button color.}
}
\description{
Add submit button
}
\examples{
library(shiny)

shinyApp(
  ui = bulmaPage(
   bulmaTitle("Hello Bulma"),
   bulmaSubmitButton("submit-button", "Submit", "primary")
  ),
  server = function(input, output) {}
)

}
