% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-input.R
\name{bulmaSwitchInput}
\alias{bulmaSwitchInput}
\title{Add switch input}
\usage{
bulmaSwitchInput(inputId, label = NULL, value = FALSE, color = NULL,
  size = NULL, style = NULL, rtl = FALSE, disabled = FALSE)
}
\arguments{
\item{label}{Switch label.}

\item{value}{TRUE or FALSE, FALSE by default.}

\item{color}{Switch color : \code{link}, \code{info}, \code{primary}, \code{warning},
\code{danger}, \code{success}, \code{black}, \code{dark} and \code{ligth}.}

\item{size}{Switch size : \code{small}, \code{medium} and \code{large}.}

\item{style}{Switch style :  \code{thin}, \code{rounded}, \code{outlined} or 
\code{mixed1}, \code{mixed2}, \code{mixed3} and \code{mixed4} (mixed1 is thin-rounded, mixed2
is thin-outlined, mixed3 is rounded-outlined and mixed4 is thin-rounded-outlined)
NULL by default, which corresponds to squared style.}

\item{inpuId}{The input slot that will be used to access the value.}
}
\description{
Create a switch input \url{https://wikiki.github.io/form/switch/}.
}
\examples{
if(interactive()) {
library(shiny)

shinyApp(
 ui = bulmaNavbarPage(
   bulmaNavbar(
     bulmaNavbarBrand(
       bulmaNavbarItem(
         "shinybulma"
       ),
       bulmaNavbarBurger()
     ),
     bulmaNavbarMenu( # not visible on smaller devices
       bulmaNavbarItem(
         "Switch Inputs"
       )
     )
   ),
   bulmaNav(
     "Switch Inputs",
     div(style = "text-align: center;",
         bulmaTitle("The plot only displays if switch 1 is TRUE.")),
     
     br(), br(), hr(),
     
     bulmaContainer(
       bulmaColumns(
         bulmaColumn(
           width = 6,
           bulmaSubtitle("Inputs"),
           bulmaSwitchInput(
             inputId = "switch1", 
             label = "switch 1", 
             value = FALSE,
             color = NULL, 
             size = NULL, 
             style = NULL,
             rtl = FALSE, 
             disabled = FALSE),
           
           bulmaSwitchInput(
             inputId = "switch2", 
             label = "switch 2", 
             value = TRUE,
             color = "warning", 
             size = "large", 
             style = "rounded",
             
             rtl = FALSE, 
             disabled = TRUE),
           
           bulmaSwitchInput(
             inputId = "switch3", 
             label = "switch 3", 
             value = TRUE,
             
             color = "danger", 
             size = "small", 
             style = "thin",
             
             rtl = FALSE, 
             disabled = FALSE),
           
           bulmaSwitchInput(
             inputId = "switch4", 
             label = "switch 4", 
             value = FALSE,
             
             color = NULL, 
             size = NULL, 
             style = "mixed3",
             rtl = FALSE, 
             disabled = FALSE)
         ),
         bulmaColumn(
           width = 6,
           bulmaSubtitle("Outputs"),
           plotOutput("plot1"),
           uiOutput("switches")
         )
       )
     )
   )
 ),
 server = function(input, output) {
   
   output$plot1 <- renderPlot({
     if (input$switch1 == TRUE) {
       plot(mtcars$wt, mtcars$mpg)
     }
   })
   
   output$switches <- renderPrint({
     c(input$switch1, input$switch2, input$switch3, input$switch4)
   })
 }
 )
}
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
