% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-tag.R
\name{bulmaTagAddon}
\alias{bulmaTagAddon}
\title{bulma tag addon}
\usage{
bulmaTagAddon(...)
}
\arguments{
\item{...}{Slot for 2 bulmaTag.}
}
\description{
Merge two tags together \url{https://bulma.io/documentation/elements/tag/}.
}
