% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-tag.R
\name{bulmaTagAddonList}
\alias{bulmaTagAddonList}
\title{bulma tag addon horizontal list.}
\usage{
bulmaTagAddonList(...)
}
\arguments{
\item{...}{Slot for bulmaTagAddons.}
}
\description{
Create an horizontal list of bulmaTagAddons objects
\url{https://bulma.io/documentation/elements/tag/}.
}
