% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-tag.R
\name{bulmaTagAddons}
\alias{bulmaTagAddons}
\title{bulma tag addon vertical list}
\usage{
bulmaTagAddons(...)
}
\arguments{
\item{...}{Slot for bulmaTagAddon.}
}
\description{
Create a vertical list of bulmaTagAddon objects
\url{https://bulma.io/documentation/elements/tag/}.
}
