% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-input.R
\name{bulmaTagInput}
\alias{bulmaTagInput}
\title{bulma TagsInput}
\usage{
bulmaTagInput()
}
\description{
Create a tag input \url{https://wikiki.github.io/form/tagsinput/}.
}
\note{
Do not need to provide any argument.
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
