% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-tag.R
\name{bulmaTagList}
\alias{bulmaTagList}
\title{bulma tag list}
\usage{
bulmaTagList(...)
}
\arguments{
\item{...}{Slot for bulmaTag.}
}
\description{
Create a tag list of bulmaTag objects \url{https://bulma.io/documentation/elements/tag/}.
}
