% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-timeline.R
\name{bulmaTimeline}
\alias{bulmaTimeline}
\title{bulma timeline}
\usage{
bulmaTimeline(..., centered = FALSE, rtl = FALSE)
}
\arguments{
\item{...}{Slot for bulmaTimelineHeader and bulmaTimelineItem.}

\item{centered}{FALSE by default. Whether to center the timeline.}

\item{rtl}{FALSE by default. If rtl is TRUE, the timeline is displayed
right to left. If FALSE (and if centered is FALSE), the timeline is displayed
left to right (by default).}
}
\description{
Create a timeline container \url{https://wikiki.github.io/components/timeline/}.
}
\note{
known problem: when centered is TRUE, lines cannot be colored.
}
\examples{
\dontrun{
if (interactive()) {
 library(shiny)

 ui <- bulmaPage(
  bulmaContainer(
   # centered timeline
   bulmaTimeline(
    centered = TRUE,
    bulmaTimelineHeader(text = "Start", size = "medium", color = "primary"),
    bulmaTimelineItem(
      color = "primary", marker_color = "primary",
      marker_image = FALSE, marker_icon = FALSE,
      content_header = "January 2016", content_body = "Timeline content - Can include any HTML element"
    ),
    bulmaTimelineItem(
      color = "warning", marker_color = "warning",
      marker_image = TRUE, marker_image_size = "32x32", marker_icon = FALSE,
      content_header = "February 2016", content_body = "Timeline content - Can include any HTML element",
      tags$img(src = "http://bulma.io/images/placeholders/32x32.png")
    ),
    bulmaTimelineHeader(text = "2017", size = NULL, color = "primary"),
    bulmaTimelineItem(
      color = "danger", marker_color = "danger",
      marker_image = FALSE, marker_icon = TRUE,
      content_header = "March 2016", content_body = "Timeline content - Can include any HTML element",
      tags$i(class = "fa fa-flag")
    ),
    bulmaTimelineHeader(text = "End", size = "medium", color = "primary")
    )
  )
 )

 server <- function(input, output, session) {

 }

 shinyApp(ui = ui, server = server)

}
}
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
