% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-timeline.R
\name{bulmaTimelineHeader}
\alias{bulmaTimelineHeader}
\title{bulma timeline header}
\usage{
bulmaTimelineHeader(text = NULL, size = NULL, color = NULL)
}
\arguments{
\item{text}{Header name.}

\item{size}{Header size : \code{small},\code{medium} and \code{large}.}

\item{color}{Header color : \code{link}, \code{info}, \code{primary}, \code{warning},
\code{danger}, \code{success}, \code{black}, \code{dark} and \code{ligth}.}
}
\description{
Create a timeline header \url{https://wikiki.github.io/components/timeline/}.
}
\note{
you can insert any number of headers.
}
