% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-timeline.R
\name{bulmaTimelineItem}
\alias{bulmaTimelineItem}
\title{bulma timeline item}
\usage{
bulmaTimelineItem(
  ...,
  color = NULL,
  marker_color = NULL,
  marker_image = FALSE,
  marker_image_size = NULL,
  marker_icon = FALSE,
  content_header = NULL,
  content_body = NULL
)
}
\arguments{
\item{...}{Insert an icon or image here (if they are enabled via marker_image and
marker_icon, respectively) : \code{tags$i(class = "fa fa-flag")} or
\code{tags$img(src = "http://bulma.io/images/placeholders/32x32.png")}}

\item{color}{Plan color : \code{link}, \code{info}, \code{primary}, \code{warning},
\code{danger}, \code{success}, \code{black}, \code{dark} and \code{ligth}.}

\item{marker_color}{Item marker color : \code{link}, \code{info}, \code{primary}, \code{warning},
\code{danger}, \code{success}, \code{black}, \code{dark} and \code{ligth}.}

\item{marker_image}{TRUE or FALSE, FALSE by default. Whether to insert an image
in the marker.}

\item{marker_image_size}{image size.}

\item{marker_icon}{TRUE or FALSE, FALSE by default. Whether to insert an icon
in the marker.}

\item{content_header}{Content header.}

\item{content_body}{Item description.}
}
\description{
Create a timeline item \url{https://wikiki.github.io/components/timeline/}.
}
\details{
Valid \code{size}:
\itemize{
  \item{16 - 16x16}
  \item{24 - 24x24}
  \item{32 - 32x32}
  \item{48 - 48x48}
  \item{54 - 54x54}
  \item{96 - 96x96}
  \item{128 - 128x128}
}
}
