% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-card.R
\name{bulmaCard}
\alias{bulmaCard}
\alias{bulmaCardHeader}
\alias{bulmaCardHeaderTitle}
\alias{bulmaCardContent}
\alias{bulmaCardFooter}
\alias{bulmaCardFooterItem}
\title{Add card}
\usage{
bulmaCard(...)

bulmaCardHeader(...)

bulmaCardHeaderTitle(...)

bulmaCardContent(...)

bulmaCardFooter(...)

bulmaCardFooterItem(..., tag = shiny::span)
}
\arguments{
\item{...}{any element.}

\item{tag}{html tag.}
}
\description{
Flexible card component.
}
\examples{
if(interactive()){
library(shiny)

shinyApp(
  ui = bulmaPage(
   bulmaSection(
     bulmaContainer(
       bulmaCard(
         bulmaCardHeader(
           bulmaCardHeaderTitle(
             "Card title"
           )
         ),
         bulmaCardContent(
           "Content of the card."
         ),
         bulmaCardFooter(
           bulmaCardFooterItem(
             "Item 1"
           ),
           bulmaCardFooterItem(
             "Item 2"
           )
         )
       )
     )
   )
  ),
  server = function(input, output) {}
)
}

}
\author{
John Coene, \email{jcoenep@ymail.com}
}
