% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-hero.R
\name{bulmaHero}
\alias{bulmaHero}
\alias{bulmaHeroHead}
\alias{bulmaHeroBody}
\alias{bulmaHeroFoot}
\title{Add section}
\usage{
bulmaHero(..., color = NULL, size = NULL, bold = FALSE,
  fullheight = FALSE)

bulmaHeroHead(...)

bulmaHeroBody(..., container = TRUE)

bulmaHeroFoot(...)
}
\arguments{
\item{...}{any element.}

\item{color}{hero color.}

\item{size}{hero size takes \code{NULL}, \code{medium} or \code{large}.}

\item{bold}{set to \code{TRUE} to generate a subtle gradient.}

\item{fullheight}{set to \code{TRUE} to set hero to \code{100vh}.}

\item{container}{If \code{TRUE} wraps a \code{\link{bulmaContainer}} as a children of the section.}
}
\description{
Add a section tag.
}
\examples{
if(interactive()){
library(shiny)

ui <- bulmaPage(
  bulmaHero(
    fullheight = TRUE,
    color = "danger",
    bulmaHeroBody(
      bulmaContainer(
        bulmaTitle("Hello Bulma*"),
        bulmaSubtitle("shinybulma*")
      )
    )
  )
)

server <- function(input, output){}

shinyApp(ui, server)
}

}
\author{
John Coene, \email{jcoenep@gmail.com}
}
