% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-levels.R
\name{bulmaLevel}
\alias{bulmaLevel}
\alias{bulmaLevelItem}
\title{Add level}
\usage{
bulmaLevel(..., mobile = FALSE)

bulmaLevelItem(heading, value, centered = TRUE)
}
\arguments{
\item{...}{any element.}

\item{mobile}{set to \code{TRUE} if you want the level to be horizontal on mobile.}
}
\description{
Add levels
}
\examples{
library(shiny)

shinyApp(
  ui = bulmaPage(
    bulmaLevel(
      bulmaLevelItem("Tweets", "3'456"),
      bulmaLevelItem("Following", 123),
      bulmaLevelItem("Followers", "456K")
    )
  ),
  server = function(input, output) {}
)

}
