% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-page.R
\name{bulmaPage}
\alias{bulmaPage}
\alias{bulmaNavbarPage}
\title{Create a bulma app}
\usage{
bulmaPage(..., title = NULL, theme = "default")

bulmaNavbarPage(..., title = NULL, theme = "default")
}
\arguments{
\item{...}{any element.}

\item{title}{page title.}

\item{theme}{a theme.}
}
\description{
Build a bulma Shiny app.
}
\section{Themes}{

You can preview all themes at \url{https://jenil.github.io/bulmaswatch/}.
\itemize{
 \item{\code{cerulean}}
 \item{\code{cyborg}}
 \item{\code{dark}}
 \item{\code{darly}}
 \item{\code{flatly}}
 \item{\code{journal}}
 \item{\code{default}}
 \item{\code{litera}}
 \item{\code{lumen}}
 \item{\code{lux}}
 \item{\code{materia}}
 \item{\code{minty}}
 \item{\code{nuclear}}
 \item{\code{pulse}}
 \item{\code{sansdstone}}
 \item{\code{simplex}}
 \item{\code{slate}}
 \item{\code{solar}}
 \item{\code{spacelab}}
 \item{\code{united}}
 \item{\code{yeti}}
}
}

\examples{
if(interactive()){
library(shiny)

shiny::shinyApp(
  ui = bulmaNavbarPage(
   bulmaNavbar(
     bulmaNavbarBrand(
       bulmaNavbarItem(
         "shinybulma"
       ),
       bulmaNavbarBurger()
     ),
     bulmaNavbarMenu( # not visible on smaller devices
       bulmaNavbarItem(
         "Item 1"
       ),
       bulmaNavbarItem(
         "Item 2"
       ),
     bulmaNavbarDropdown(
       label = "Drop-down",
       bulmaNavbarItem(
         "Select 1"
       ),
       bulmaNavbarItem(
         "Select 2"
       )
     )
    )
   ),
   bulmaNav(
     "Item 1",
     bulmaTitle("Content for item 1 here.")
   ),
   bulmaNav(
     "Item 2",
     bulmaTitle("Content for item 2 here.")
   )
  ),
  server = function(input, output) {}
)
}

}
\author{
John Coene, \email{jcoenep@gmail.com}
}
