% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-progress.R
\name{bulmaProgress}
\alias{bulmaProgress}
\alias{renderBulmaProgress}
\alias{bulmaProgressOutput}
\title{Add a progress bar}
\usage{
bulmaProgress(..., value, max = 100, color = NULL, size = NULL)

renderBulmaProgress(expr, env = parent.frame(), quoted = FALSE)

bulmaProgressOutput(outputId)
}
\arguments{
\item{...}{any element.}

\item{value}{value of progress.}

\item{max}{maximum value.}

\item{color}{color of progress bar.}

\item{size}{size of progress bar, default to \code{NULL}
takes \code{small}, \code{meidum} or \code{large}.}

\item{expr}{An expression that returns a Shiny tag object, \code{\link[shiny:reexports]{HTML()}},
or a list of such objects.}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}

\item{outputId}{id of element.}
}
\description{
Add a progress bar.
}
\examples{
if(interactive()){
library(shiny)

shinyApp(
  ui = bulmaPage(
   bulmaTitle("Hello Bulma"),
   bulmaProgress("15\%", value = 15, max = 100),
   bulmaProgress("25\%", value = 25, max = 100, color = "primary"),
   bulmaProgress("65\%", value = 65, max = 100, color = "danger", size = "medium"),
   bulmaProgressOutput("progress")
  ),
  server = function(input, output) {
    output$progress <- renderBulmaProgress({
      bulmaProgress(
        value = sample(1:100, 1)
      )
    })
  }
)
}

}
