% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-tabs.R
\name{bulmaTabs}
\alias{bulmaTabs}
\alias{bulmaTab}
\title{Add tabs}
\usage{
bulmaTabs(tabs, center = FALSE, ...)

bulmaTab(label, ...)
}
\description{
Add tabs

Add tab
}
\examples{
library(shiny)

shinyApp(
  ui = bulmaPage(
   bulmaTitle("Hello Bulma"),
   bulmaTabs(
     tabs = c("Tab 1", "Tab 2", "Tab 3"),
     center = TRUE,
     bulmaTab(
       "Tab 1",
       "Content of the first Tab."
     ),
     bulmaTab(
       "Tab 2",
       plotOutput("plot")
     ),
     bulmaTab(
       "Tab 3",
       "Content of the third Tab."
     )
   )
  ),
  server = function(input, output) {
    output$plot <- renderPlot({
      hist(rnorm(20, 10))
    })
  }
)

}
