% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-tiles.R
\name{bulmaTileAncestor}
\alias{bulmaTileAncestor}
\alias{bulmaTileParent}
\alias{bulmaTileChild}
\title{Add tiles}
\usage{
bulmaTileAncestor(...)

bulmaTileParent(..., vertical = FALSE, width = NULL)

bulmaTileChild(..., title = "", subtitle = "", color = NULL)
}
\arguments{
\item{...}{any element.}

\item{vertical}{set to \code{TRUE} if you want to stack tiles vertically.}

\item{width}{define width of tile.}
}
\description{
Add masonry-like tiles.
}
\examples{
library(shiny)

shinyApp(
  ui = bulmaPage(
   bulmaTileAncestor(
     bulmaTileParent(
       vertical = TRUE,
       bulmaTileChild(
         bulmaTitle("Tile 1"),
         tags$p("some graph")
       ),
       bulmaTileChild(
         bulmaTitle("Tile 2"),
         plotOutput("graph")
       )
     ),
     bulmaTileParent(
       bulmaTileChild(
         bulmaTitle("Tile 3"),
         tags$p("some other graph")
       ),
       bulmaTileChild(
         bulmaTitle("Tile 5")
       ),
       bulmaTileChild(
         bulmaTitle("Tile 6")
       )
     )
   )
  ),
  server = function(input, output) {
    output$graph <- renderPlot({
      plot(x = runif(20, 5, 10), y = runif(20, 10, 12))
    })
  }
)


}
