% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-title.R
\name{bulmaTitle}
\alias{bulmaTitle}
\alias{bulmaSubtitle}
\title{Add subtitle}
\usage{
bulmaTitle(...)

bulmaSubtitle(...)
}
\arguments{
\item{...}{any element.}
}
\description{
Add subtitle
}
\examples{
shinyApp(
  ui = bulmaPage(
    bulmaHero(
      fullheight = TRUE,
      color = "primary",
      bulmaHeroHead(
        bulmaContainer(
          bulmaTitle("Hero Title")
        )
      ),
      bulmaHeroHead(
        bulmaContainer(
          bulmaSubtitle(
            "Some content"
          )
        )
      ),
      bulmaHeroFoot(
        bulmaContainer(
          "The footer"
        )
      )
    )
  ),
  server = function(input, output) {}
)

}
