% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulmaBtn.R
\name{UpdateBulmaBtn}
\alias{UpdateBulmaBtn}
\title{Update Bulma Button}
\usage{
UpdateBulmaBtn(
  inputId,
  text = NULL,
  color = NULL,
  size = NULL,
  state = NULL,
  spacing = NULL,
  responsive = NULL,
  fullwidth = NULL,
  outlined = NULL,
  inverted = NULL,
  rounded = NULL,
  disabled = NULL,
  hidden = NULL,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{inputId}{string value for id}

\item{text}{NULL or string value for button text}

\item{color}{NULL or string value for button color}

\item{size}{NULL or string value for button size}

\item{state}{NULL or string value for button state}

\item{spacing}{NULL or string value for margin and padding}

\item{responsive}{NULL or Boolean value for responsive size}

\item{fullwidth}{NULL or Boolean value to make button size fullwidth}

\item{outlined}{NULL or Boolean value to make button outlined}

\item{inverted}{NULL or Boolean value to make button inverted}

\item{rounded}{NULL or Boolean value to make button with rounded corners}

\item{disabled}{NULL or Boolean value to make a button disabled}

\item{hidden}{NULL or Boolean value to make a button hidden}

\item{session}{session object}
}
\description{
Update Bulma Button
}
