% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-input.R
\name{bulmaActionButton}
\alias{bulmaActionButton}
\title{Add submit button}
\usage{
bulmaActionButton(inputId, label, color = NULL)
}
\arguments{
\item{inputId}{the input slot that will be used to access the value.}

\item{label}{button label.}

\item{color}{button color.}
}
\description{
Add submit button
}
\examples{
if(interactive()){
library(shiny)

shinyApp(
  ui = bulmaPage(
    bulmaActionButton("go", "Go")
  ),
  server = function(input, output, session) {
    observeEvent(input$go, {
      print(paste("This will only be printed once; all",
        "subsequent button clicks won't do anything"))
    }, once = TRUE)
 }
)
}
}
