% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulmaBtn.R
\name{BulmaBtn}
\alias{BulmaBtn}
\title{Bulma Button}
\usage{
BulmaBtn(
  inputId,
  text,
  size = "is-normal",
  color = NULL,
  state = NULL,
  responsive = FALSE,
  fullwidth = FALSE,
  outlined = FALSE,
  inverted = FALSE,
  rounded = FALSE,
  disabled = FALSE,
  hidden = FALSE
)
}
\arguments{
\item{inputId}{string value for the id}

\item{text}{string value for the text displayed in the button}

\item{size}{string value for size; choices are: "is-normal", "is-small",
"is-medium", "is-large"}

\item{color}{string value for color; choices are: "is-white", "is-light",
"is-dark", "is-black", "is-text", "is-ghost", "is-primary", "is-link",
"is-info", "is-success", "is-warning", "is-danger"}

\item{state}{string value for state; choices are: "is-hovered",
"is-focused", "is-active", "is-loading"}

\item{responsive}{if true, will make the button size responsive to the
window size}

\item{fullwidth}{if true, will increase the button width to fit its parent
container}

\item{outlined}{if true, will create an outlined design}

\item{inverted}{if true, will create an inverted design}

\item{rounded}{if true, will create a rounded button}

\item{disabled}{if true, will disable the button}

\item{hidden}{if true, will hide the button from view}
}
\description{
Bulma Button
}
\examples{
BulmaBtn(inputId = "example", text = "Submit")
}
