% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-container.R
\name{bulmaContainer}
\alias{bulmaContainer}
\title{Add container}
\usage{
bulmaContainer(..., fluid = FALSE, widescreen = FALSE, fullhd = FALSE)
}
\arguments{
\item{...}{any element.}

\item{fluid}{If you don't want to have a maximum width but want to keep the 24px margin on the left and right sides.}

\item{fullhd}{until the \code{fullhd} breakpoint.}

\item{fullwidth}{until the \code{widescreen} breakpoint.}
}
\description{
Add container
}
