% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-input.R
\name{bulmaDateInput}
\alias{bulmaDateInput}
\title{Date picker}
\usage{
bulmaDateInput(inputId, value, min = Sys.Date() - 3, max = Sys.Date(),
  overlay = FALSE, lang = "en", class = NULL, ...)
}
\arguments{
\item{inputId}{Id to access value.}

\item{value}{Current date.}

\item{min, max}{Minimum and maximum selectable dates.}

\item{overlay}{Set to \code{TRUE} to heave the calendar appear as overlay.}

\item{lang}{Language.}

\item{class}{Additional CSS class.}

\item{...}{any other option}
}
\description{
Add date picker.
}
\note{
Does not work in RStudio, open app in browser.
}
\examples{
if(interactive()){
  ui <- bulmaPage(
    bulmaContainer(
      br(),
      bulmaDateInput("date", Sys.Date(), overlay = TRUE),
      verbatimTextOutput("selected")
    )
  )
  
  server <- function(input, output){
    
    output$selected <- renderText({
      input$date
    })
  }
  
  shiny::shinyApp(ui, server)
}

}
\author{
John Coene, \email{jcoenep@gmail.com}
}
