% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-footer.R
\name{bulmaFooter}
\alias{bulmaFooter}
\title{Add a footer}
\usage{
bulmaFooter(..., container = TRUE)
}
\arguments{
\item{...}{any element.}
}
\description{
Add a footer
}
\examples{
library(shiny)
shinyApp(
  ui = bulmaPage(
    bulmaFooter(
      shiny::p(
        "The footer"
      )
    )
  ),
  server = function(input, output) {}
)

}
