% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-modal.R
\name{bulmaModal}
\alias{bulmaModal}
\title{Modal}
\usage{
bulmaModal(id, options, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{Unique modal id.}

\item{options}{Modal configuration. See \url{https://github.com/VizuaaLOG/BulmaJS}.}

\item{session}{A valid shiny session.}
}
\description{
Display a modal
}
\examples{
library(shiny)
ui <- bulmaPage(
 bulmaActionButton("show_modal", "Show modal"),
 bulmaActionButton("show_modal2", "Show modal 2")
)

server <- function(input, output){
 observeEvent(input$show_modal, {
   bulmaModal(
    id = "mymodal", 
    list(
     title = "Modal title",
     body = "Modal content"
    )
   )
 })
 observeEvent(input$show_modal2, {
   bulmaModal(
    id = "mymodal2", 
    list(
     title = "Modal title 2",
     body = "Modal content 2"
    )
   )
 })
}

if(interactive())
 shinyApp(ui, server)

}
