% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-notification.R
\name{bulmaNotification}
\alias{bulmaNotification}
\title{bulma notification}
\usage{
bulmaNotification(
  id = NULL,
  options = NULL,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{id}{Notification target. It will be attached on the provided DOM element. If id is NULL,
the notification is attached to the body.}

\item{options}{A list of notification options. See \url{https://github.com/VizuaaLOG/BulmaJS}.}

\item{session}{A valid shiny session.}
}
\description{
Create a notification box \url{https://bulma.io/documentation/elements/notification/}.
}
\examples{
if (interactive()) {
 library(shiny)

 ui <- bulmaPage(
  bulmaActionButton("show_notif", "Show notification"),
  div(id = "target", style = "top: 0px; right: 0px; position: absolute;")
 )
 
 server <- function(input, output, session) {
  observeEvent(input$show_notif, {
    bulmaNotification(
      id = "target", 
      options = list(
        body = "This is the message",
        color = "info",
        dismissInterval = 200000,
        isDismissable = TRUE
      )
    )
  })
 }
 shinyApp(ui, server)
}
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
