% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-notification.R
\name{bulmaNotification}
\alias{bulmaNotification}
\title{bulma notification}
\usage{
bulmaNotification(..., color = NULL)
}
\arguments{
\item{...}{Any element.}

\item{color}{The background color : \code{link}, \code{info}, \code{primary}, \code{warning},
\code{danger}, \code{success}, \code{black}, \code{dark} and \code{ligth}.}
}
\description{
Create a notification box \url{https://bulma.io/documentation/elements/notification/}.
}
\examples{
\dontrun{
if (interactive()) {
 library(shiny)

 ui <- bulmaPage(
  bulmaContainer(
   bulmaColumns(
    bulmaColumn(
      width = 3,
      bulmaNotification(
        color = "info",
        "Lorem ipsum dolor sit amet, consectetur adipiscing
        elit lorem ipsum dolor. Pellentesque risus mi, tempus
        quis placerat ut, porta nec nulla. Vestibulum rhoncus
        ac ex sit amet fringilla. Nullam gravida purus diam,
        et dictum felis venenatis efficitur. Sit amet,
        consectetur adipiscing elit"
      )
    ),
    bulmaColumn(
      width = 3,
      bulmaNotification(
        color = "warning",
        "Lorem ipsum dolor sit amet, consectetur adipiscing
        elit lorem ipsum dolor. Pellentesque risus mi, tempus
        quis placerat ut, porta nec nulla. Vestibulum rhoncus
        ac ex sit amet fringilla. Nullam gravida purus diam,
        et dictum felis venenatis efficitur. Sit amet,
        consectetur adipiscing elit"
      )
    ),
    bulmaColumn(
      width = 3,
      bulmaNotification(
        color = "danger",
        "Lorem ipsum dolor sit amet, consectetur adipiscing
        elit lorem ipsum dolor. Pellentesque risus mi, tempus
        quis placerat ut, porta nec nulla. Vestibulum rhoncus
        ac ex sit amet fringilla. Nullam gravida purus diam,
        et dictum felis venenatis efficitur. Sit amet,
        consectetur adipiscing elit"
      )
    ),
    bulmaColumn(
      width = 3,
      bulmaNotification(
        color = "primary",
        "Lorem ipsum dolor sit amet, consectetur adipiscing
        elit lorem ipsum dolor. Pellentesque risus mi, tempus
        quis placerat ut, porta nec nulla. Vestibulum rhoncus
        ac ex sit amet fringilla. Nullam gravida purus diam,
        et dictum felis venenatis efficitur. Sit amet,
        consectetur adipiscing elit"
      )
    )
   )
  )
 )

 server <- function(input, output, session) {

 }

 shinyApp(ui = ui, server = server)

}
}
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
