% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-input.R
\name{bulmaRadioInput}
\alias{bulmaRadioInput}
\title{Add radio buttons}
\usage{
bulmaRadioInput(inputId, choices, selected)
}
\arguments{
\item{inputId}{the input slot that will be used to access the value.}

\item{choices}{vector of choices, named or unnamed.}

\item{selected}{selected radio input.}
}
\description{
Add radio buttons
}
\examples{
if(interactive()){
library(shiny)

shinyApp(
  ui = bulmaPage(
   bulmaTitle("Hello Bulma"),
   bulmaRadioInput("select", c("Miles per galon" = "mpg", "Rear axle ratio" = "drat"),
     selected = "mpg"),
   plotOutput("plot")
  ),
  server = function(input, output) {
    output$plot <- renderPlot(
      plot(1:nrow(mtcars), mtcars[[input$select]])
    )
  }
)
}

}
