% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-section.R
\name{bulmaSection}
\alias{bulmaSection}
\title{Add a section}
\usage{
bulmaSection(..., size = NULL)
}
\arguments{
\item{...}{any element.}

\item{size}{change the spacing, takes \code{medium} or \code{large}.}
}
\description{
Add a container to divide your page into sections
}
\examples{
library(shiny)

shinyApp(
  ui = bulmaPage(
   bulmaSection(
     size = "large",
     bulmaContainer(
      bulmaTitle("Section 1 title"),
      bulmaSubtitle("Section 1 subtitle")
     )
   ),
   bulmaSection(
     bulmaContainer(
      bulmaTitle("Section 1 title"),
      bulmaSubtitle("Section 1 subtitle")
     )
   )
  ),
  server = function(input, output) {}
)

}
