% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-levels.R
\name{bulmaLevel}
\alias{bulmaLevel}
\alias{bulmaLevelItem}
\alias{renderBulmaLevel}
\alias{bulmaLevelOutput}
\title{Add level}
\usage{
bulmaLevel(..., mobile = FALSE)

bulmaLevelItem(heading, value, centered = TRUE)

renderBulmaLevel(expr, env = parent.frame(), quoted = FALSE)

bulmaLevelOutput(outputId)
}
\arguments{
\item{...}{any element.}

\item{mobile}{set to \code{TRUE} if you want the level to be horizontal on mobile.}

\item{heading}{level heading.}

\item{value}{level value.}

\item{centered}{whether to center to level.}

\item{expr}{An expression that returns a Shiny tag object, \code{\link[shiny:reexports]{HTML()}},
or a list of such objects.}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}

\item{outputId}{id of output.}
}
\description{
Add levels
}
\examples{
if(interactive()){
library(shiny)

shinyApp(
  ui = bulmaPage(
    bulmaLevel(
      bulmaLevelItem("Tweets", "3'456"),
      bulmaLevelItem("Following", 765),
      bulmaLevelItem("Followers", "2'134")
    ),
    bulmaLevelOutput("levels")
  ),
  server = function(input, output) {
    output$levels <- renderBulmaLevel({
      bulmaLevel(
        bulmaLevelItem("Tweets", sample(100:5000, 1)),
        bulmaLevelItem("Following", sample(100:5000, 1)),
        bulmaLevelItem("Followers", sample(100:5000, 1))
      )
    })
  }
)
}

}
\author{
John Coene, \email{jcoenep@gmail.com}
}
