% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-navbar.R
\name{bulmaNavbar}
\alias{bulmaNavbar}
\alias{bulmaNavbarBrand}
\alias{bulmaNavbarMenu}
\alias{bulmaNavbarItem}
\alias{bulmaNavbarLink}
\alias{bulmaNavbarBurger}
\alias{bulmaNavbarDropdown}
\title{Add navbar}
\usage{
bulmaNavbar(..., transparent = FALSE, color = NULL, fix.top = FALSE,
  fix.bottom = FALSE)

bulmaNavbarBrand(...)

bulmaNavbarMenu(...)

bulmaNavbarItem(..., href = "")

bulmaNavbarLink(..., href = "")

bulmaNavbarBurger()

bulmaNavbarDropdown(..., label, href = "")
}
\arguments{
\item{...}{any element.}
}
\description{
Add a responsive horizontal navbar.
}
\examples{
library(shiny)

shinyApp(
  ui = bulmaPage(
   bulmaNavbar(
     fix.top = TRUE,
     color = "primary",
     bulmaNavbarBrand(
       bulmaNavbarItem(
         "shinybulma"
       ),
       bulmaNavbarBurger()
     ),
     bulmaNavbarMenu( # not visible on smaller devices
       bulmaNavbarItem(
         "Item  1"
       ),
       bulmaNavbarItem(
         "Item 2"
       ),
     bulmaNavbarDropdown(
       label = "Drop-down",
       bulmaNavbarItem(
         "Select 1"
       ),
       bulmaNavbarItem(
         "Select 2"
       )
     )
     )
   )
  ),
  server = function(input, output) {}
)

}
