% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-navbar.R
\name{bulmaNavbar}
\alias{bulmaNavbar}
\alias{bulmaNavbarBrand}
\alias{bulmaNavbarMenu}
\alias{bulmaNavbarItem}
\alias{bulmaNavbarLink}
\alias{bulmaNavbarBurger}
\alias{bulmaNavbarDropdown}
\alias{bulmaNav}
\title{Add navbar}
\usage{
bulmaNavbar(..., transparent = FALSE, color = NULL, fix.top = FALSE,
  fix.bottom = FALSE)

bulmaNavbarBrand(...)

bulmaNavbarMenu(...)

bulmaNavbarItem(label, href = NULL)

bulmaNavbarLink(label, href = "")

bulmaNavbarBurger()

bulmaNavbarDropdown(..., label, href = "")

bulmaNav(target, ...)
}
\arguments{
\item{...}{any element.}

\item{transparent}{set to have transparent.}

\item{color}{navbar color.}

\item{fix.top, fix.bottom}{set one to have navbar fixed at top or bottom of page.}

\item{label}{item label.}

\item{href, target}{id of \code{bulmaNav} linked, \code{href} and \code{target} must be identical.}
}
\description{
Add a responsive horizontal navbar.
}
\examples{
if(interactive()){
library(shiny)
shinyApp(
  ui = bulmaPage(
   bulmaNavbar(
     bulmaNavbarBrand(
       bulmaNavbarItem(
         "shinybulma",
         href = "Item 1"
       ),
       bulmaNavbarBurger()
     ),
     bulmaNavbarMenu( # not visible on smaller devices
       bulmaNavbarItem(
         "Item 1"
       ),
       bulmaNavbarItem(
         "Item 2"
       ),
       bulmaNavbarItem(
         "Item 3"
       )
    )
   ),
   bulmaNav(
     "Item 1",
     bulmaContainer(
       bulmaTitle("Item 1"),
       bulmaTabs(
       tabs = c("Tab 1", "Tab 2", "Tab 3"),
       center = TRUE,
         bulmaTab(
           "Tab 1",
           bulmaTitle("Tab 1")
         ),
         bulmaTab(
           "Tab 2",
           bulmaTitle("Tab 2"),
           plotOutput("hist")
         ),
         bulmaTab(
           "Tab 3",
           bulmaTitle("Tab 3")
         )
       )
     )
   ),
   bulmaNav(
     "Item 2",
     bulmaContainer(
       bulmaTitle("Item 2"),
       plotOutput("plot2"),
       bulmaTabs(
       tabs = c("Tab 1", "Tab 2", "Tab 3"),
       center = TRUE,
         bulmaTab(
           "Tab 1",
           bulmaTitle("Tab 1")
         ),
         bulmaTab(
           "Tab 2",
           bulmaTitle("Tab 2"),
           plotOutput("plot")
         ),
         bulmaTab(
           "Tab 3",
           bulmaTitle("Tab 3")
         )
       )
     )
   ),
   bulmaNav(
     "Item 3",
     bulmaContainer(
       bulmaTitle("Item 3")
     )
   )
  ),
  server = function(input, output) {
    output$hist <- renderPlot({
      hist(rnorm(20, 10))
    })
    output$plot <- renderPlot({
      plot(1:20, rnorm(20, 20))
    })
    
    output$plot2 <- renderPlot({
      plot(1:20, rnorm(20, 20))
    })
  }
)
}

}
\author{
John Coene, \email{jcoenep@gmail.com}
}
