% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-navbar.R
\name{bulmaNavbar}
\alias{bulmaNavbar}
\alias{bulmaNavbarBrand}
\alias{bulmaNavbarMenu}
\alias{bulmaNavbarItem}
\alias{bulmaNavbarLink}
\alias{bulmaNavbarBurger}
\alias{bulmaNavbarDropdown}
\alias{bulmaNav}
\title{Add navbar}
\usage{
bulmaNavbar(..., transparent = FALSE, color = NULL, fix.top = FALSE,
  fix.bottom = FALSE)

bulmaNavbarBrand(...)

bulmaNavbarMenu(...)

bulmaNavbarItem(label, href = NULL)

bulmaNavbarLink(label, href = "")

bulmaNavbarBurger()

bulmaNavbarDropdown(..., label, href = "")

bulmaNav(target, ...)
}
\arguments{
\item{...}{any element.}
}
\description{
Add a responsive horizontal navbar.
}
\examples{
library(shiny)

shinyApp(
  ui = bulmaPage(
   bulmaNavbar(
     bulmaNavbarBrand(
       bulmaNavbarItem(
         "shinybulma"
       ),
       bulmaNavbarBurger()
     ),
     bulmaNavbarMenu( # not visible on smaller devices
       bulmaNavbarItem(
         "Item 1"
       ),
       bulmaNavbarItem(
         "Item 2"
       ),
     bulmaNavbarDropdown(
       label = "Drop-down",
       bulmaNavbarItem(
         "Select 1"
       ),
       bulmaNavbarItem(
         "Select 2"
       )
     )
    )
   ),
   bulmaNav(
     "Item 1",
     bulmaContainer(
       bulmaTitle("Item 1")
     )
   ),
   bulmaNav(
     "Item 2",
     bulmaContainer(
       bulmaTitle("Item 2")
     )
   )
  ),
  server = function(input, output) {}
)

}
