% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-page.R
\name{bulmaPage}
\alias{bulmaPage}
\alias{bulmaNavbarPage}
\title{Create a bulma app}
\usage{
bulmaPage(...)

bulmaNavbarPage(...)
}
\arguments{
\item{...}{any element.}
}
\description{
Build a bulma Shiny app.
}
\examples{
library(shiny)

shinyApp(
  ui = bulmaNavbarPage(
   bulmaNavbar(
     bulmaNavbarBrand(
       bulmaNavbarItem(
         "shinybulma"
       ),
       bulmaNavbarBurger()
     ),
     bulmaNavbarMenu( # not visible on smaller devices
       bulmaNavbarItem(
         "Item 1"
       ),
       bulmaNavbarItem(
         "Item 2"
       ),
     bulmaNavbarDropdown(
       label = "Drop-down",
       bulmaNavbarItem(
         "Select 1"
       ),
       bulmaNavbarItem(
         "Select 2"
       )
     )
    )
   ),
   bulmaNav(
     "Item 1",
     bulmaTitle("Content for item 1 here.")
   ),
   bulmaNav(
     "Item 2",
     bulmaTitle("Content for item 2 here.")
   )
  ),
  server = function(input, output) {}
)

}
