% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-tabs.R
\name{bulmaTabs}
\alias{bulmaTabs}
\alias{bulmaTab}
\title{Add tabs}
\usage{
bulmaTabs(tabs, center = FALSE, ...)

bulmaTab(label, ...)
}
\arguments{
\item{tabs}{\code{bulmaTab} elements.}

\item{center}{set to center.}

\item{...}{any element.}

\item{label}{label of tab.}
}
\description{
Add tabs

Add tab
}
\examples{
if(interactive()){
library(shiny)

shinyApp(
  ui = bulmaPage(
   bulmaTitle("Hello Bulma"),
   bulmaTabs(
     tabs = c("Tab 1", "Tab 2", "Tab 3"),
     center = TRUE,
     bulmaTab(
       "Tab 1",
       "Content of the first Tab."
     ),
     bulmaTab(
       "Tab 2",
       plotOutput("hist")
     ),
     bulmaTab(
       "Tab 3",
       bulmaContainer(
         bulmaTitle("Plot"),
         plotOutput("plot")
       )
     )
   )
  ),
  server = function(input, output) {
    output$hist <- renderPlot({
      hist(rnorm(20, 10))
    })
    output$plot <- renderPlot({
      plot(1:20, rnorm(20, 20))
    })
  }
)
}

}
