% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-tiles.R
\name{bulmaTileAncestor}
\alias{bulmaTileAncestor}
\alias{bulmaTileParent}
\alias{bulmaTileChild}
\title{Add tiles}
\usage{
bulmaTileAncestor(...)

bulmaTileParent(..., vertical = FALSE, width = NULL)

bulmaTileChild(
  ...,
  title = "",
  subtitle = "",
  color = NULL,
  type = "notification"
)
}
\arguments{
\item{...}{any element.}

\item{vertical}{set to \code{TRUE} if you want to stack tiles vertically.}

\item{width}{define width of tile.}

\item{title, subtitle}{tile title and subtitle.}

\item{color}{color of tile.}

\item{type}{type of tile: notification or box.}
}
\description{
Add masonry-like tiles.
}
\examples{
if(interactive()){
library(shiny)

shinyApp(
  ui = bulmaPage(
   bulmaTileAncestor(
     bulmaTileParent(
       vertical = TRUE,
       bulmaTileChild(
         bulmaTitle("Tile 1"),
         tags$p("some graph"),
         color = "primary"
       ),
       bulmaTileChild(
         bulmaTitle("Tile 2"),
         plotOutput("graph")
       )
     ),
     bulmaTileParent(
       bulmaTileChild(
         bulmaTitle("Tile 3"),
         tags$p("some other graph")
       ),
       bulmaTileChild(
         bulmaTitle("Tile 5"),
         tags$p("some other graph")
       ),
       bulmaTileChild(
         bulmaTitle("Tile 6"),
         tags$p("some other graph")
       )
     )
   )
  ),
  server = function(input, output) {
    output$graph <- renderPlot({
      plot(x = runif(20, 5, 10), y = runif(20, 10, 12))
    })
  }
)
}

}
\author{
John Coene, \email{jcoenep@gmail.com}
}
