% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulma-title.R
\name{bulmaTitle}
\alias{bulmaTitle}
\alias{bulmaSubtitle}
\title{Add subtitle}
\usage{
bulmaTitle(..., class = NULL, tag = shiny::h1, id = NULL)

bulmaSubtitle(..., class = NULL, tag = shiny::h2, id = NULL)
}
\arguments{
\item{...}{any element.}

\item{class}{A valid \code{CSS} class.}

\item{tag}{html tag.}

\item{id}{target to set an html anchor.}
}
\description{
Add subtitle
}
\examples{
if(interactive()){
shinyApp(
  ui = bulmaPage(
    bulmaHero(
      fullheight = TRUE,
      color = "primary",
      bulmaHeroHead(
        bulmaContainer(
          bulmaTitle("Hero Title")
        )
      ),
      bulmaHeroHead(
        bulmaContainer(
          bulmaSubtitle(
            "Some content"
          )
        )
      ),
      bulmaHeroFoot(
        bulmaContainer(
          "The footer"
        )
      )
    )
  ),
  server = function(input, output) {}
)
}

}
\author{
John Coene, \email{jcoenep@gmail.com}
}
